Added a config class
For using config after plugin onEnable.

Doing it this way improves performance as a whole but uses just a little bit more RAM. This is because before each time we needed some data from the config we got a new instance of the config. Now we store all data from the config in this class which you can get. On reload the values get reset.